+++ data/office-data/src/main/java/com/clip/OfficeDataApplication.java
@@ -1,4 +1,4 @@
-package com.clip.officedata;
+package com.clip;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
